step18 code review fix4 create error handling, ja.yml, views

+++ todo/app/controllers/projects_controller.rb
@@ -3,12 +3,11 @@
 class ProjectsController < ApplicationController
   before_action :set_project, only: %i[show edit update destroy]
   before_action :logged_in_user
-  before_action :check_userpj_auth, only: :destroy
-  include UserProjectsHelper
 
   def index
     @projects = Project.all
     @status_list = Project.statuses.keys
+    @user_have_pj = current_user.projects.ids
   end
 
   def show
@@ -21,9 +20,13 @@ class ProjectsController < ApplicationController
   def create
     @project = Project.new(project_params)
     if @project.save
-      @project.users << @current_user
-      flash[:notice] = I18n.t('flash.succeeded', model: 'プロジェクト', action: '作成')
-      redirect_to projects_url
+      if add_user_to_project(user)
+        flash[:notice] = I18n.t('flash.succeeded', model: 'プロジェクト', action: '作成')
+        redirect_to projects_url
+      else
+        flash[:error] = I18n.t('flash.failed', model: 'ユーザプロジェクト', action: '作成')
+        redirect_to projects_url
+      end
     else
       flash.now[:error] = I18n.t('flash.failed', model: 'プロジェクト', action: '作成')
       render :new
@@ -44,7 +47,7 @@ class ProjectsController < ApplicationController
   end
 
   def destroy
-    check_userpj_auth
+    return unless current_user.projects.include?(@project)
     if @project.destroy
       flash[:notice] = I18n.t('flash.succeeded', model: 'プロジェクト', action: '削除')
       redirect_to projects_url
@@ -54,14 +57,20 @@ class ProjectsController < ApplicationController
     end
   end
 
-  def set_project
-    @project = Project.find(params[:id])
+  private
+
+  def add_user_to_project(user)
+    begin
+      return if user.projects.ids.include?(user)
+      @project.users << current_user
+    rescue
+      return false
+    end
+    true
   end
 
-  def check_userpj_auth
-    return if @current_user.projects.ids.include?(@project.id)
-    flash.now[:error] = I18n.t('flash.failed', model: 'プロジェクト', action: '削除')
-    redirect_to projects_url
+  def set_project
+    @project = Project.find(params[:id])
   end
 
   def project_params
